Add age filter, reorganize types and utils, add tests for age filter
