+++ src/components/Search/Filters/DistanceInput.tsx
@@ -3,7 +3,8 @@ import { Dispatch, SetStateAction } from "react";
 import { useTranslation } from "react-i18next";
 import { AnalyticsAction, logEvent } from "../../../analytics";
 import { SearchFilters } from "../../../types";
-import { MILE_DISTANCE_OPTIONS } from "../../../util";
+import { MILE_DISTANCE_OPTIONS } from "../../../utils";
+import FilterRadio from "./FilterRadio";
 
 type DistanceInputProps = {
   hideLegend?: boolean;
@@ -29,27 +30,21 @@ function DistanceInput({
     setFilters({ ...filters, miles });
   };
 
-  // TODO: consolidate getRadio logic? (see HelpRecipientInput)
-  const getRadio = (miles: string) => (
-    <Radio
-      id={miles.toString()}
-      name="distance"
-      label={t(`${tPrefix}withinMiles`, {
-        n: miles,
-      })}
-      checked={filters.miles === miles}
-      onChange={() => setDistanceFilter(miles)}
-      value={miles}
-      key={miles}
-    />
-  );
-
   return (
     <Fieldset
       legend={t(`${tPrefix}distance`)}
       legendStyle={hideLegend ? "srOnly" : "large"}
     >
-      {MILE_DISTANCE_OPTIONS.map((miles) => getRadio(miles))}
+      {MILE_DISTANCE_OPTIONS.map((miles) => (
+        <FilterRadio
+          name="distance"
+          value={miles}
+          label={t(`${tPrefix}withinMiles`, { n: miles })}
+          selected={filters.miles === miles}
+          onChange={() => setDistanceFilter(miles)}
+          key={miles}
+        />
+      ))}
     </Fieldset>
   );
 }
