+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./search";
+export * from "./care_provider";
